rename list to explore
